+++ packages/expo-image-picker/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### 💡 Others
 
+## 15.0.5 — 2024-05-15
+
+### 🐛 Bug fixes
+
+- [Android] Add support for granular permissions. ([#28897](https://github.com/expo/expo/pull/28897) by [@lukmccall](https://github.com/lukmccall))
+
 ## 15.0.4 — 2024-05-01
 
 _This version does not introduce any user-facing changes._